fix: resolve type conversion errors in base64 encoding
Fix bitwise operations in _encode function by properly casting bytes1 to uint8 before operations. Resolves Solidity 0.8.24 compilation errors.

+++ src/InvoiceNFT.sol
@@ -110,18 +110,18 @@ contract InvoiceNFT is ERC721Upgradeable, OwnableUpgradeable, UUPSUpgradeable {
         uint256 j;
         for (i = 0; i < len - 2; i += 3) {
             result[j++] = table[uint8(data[i] >> 2)];
-            result[j++] = table[uint8(((data[i] & 0x03) << 4) | (uint8(data[i + 1]) >> 4))];
-            result[j++] = table[uint8(((data[i + 1] & 0x0f) << 2) | (uint8(data[i + 2]) >> 6))];
+            result[j++] = table[uint8((uint8(data[i] & 0x03) << 4) | (uint8(data[i + 1]) >> 4))];
+            result[j++] = table[uint8((uint8(data[i + 1] & 0x0f) << 2) | (uint8(data[i + 2]) >> 6))];
             result[j++] = table[uint8(data[i + 2] & 0x3f)];
         }
 
         if (len % 3 == 1) {
             result[j++] = table[uint8(data[len - 1] >> 2)];
-            result[j++] = table[uint8((data[len - 1] & 0x03) << 4)];
+            result[j++] = table[uint8(uint8(data[len - 1] & 0x03) << 4)];
         } else if (len % 3 == 2) {
             result[j++] = table[uint8(data[len - 2] >> 2)];
-            result[j++] = table[uint8(((data[len - 2] & 0x03) << 4) | (uint8(data[len - 1]) >> 4))];
-            result[j++] = table[uint8((data[len - 1] & 0x0f) << 2)];
+            result[j++] = table[uint8((uint8(data[len - 2] & 0x03) << 4) | (uint8(data[len - 1]) >> 4))];
+            result[j++] = table[uint8(uint8(data[len - 1] & 0x0f) << 2)];
         }
 
         return string(result);
